load web reader container from read now button action
