+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/reader/WebReader.kt
@@ -41,123 +41,6 @@ import kotlinx.coroutines.launch
 import java.util.*
 import kotlin.math.roundToInt
 
-
-@Composable
-fun WebReaderLoadingContainer(slug: String, webReaderViewModel: WebReaderViewModel) {
-  val onBackPressedDispatcher = LocalOnBackPressedDispatcherOwner.current?.onBackPressedDispatcher
-
-  var isMenuExpanded by remember { mutableStateOf(false) }
-  var showWebPreferencesDialog by remember { mutableStateOf(false ) }
-
-  val webReaderParams: WebReaderParams? by webReaderViewModel.webReaderParamsLiveData.observeAsState(null)
-  val annotation: String? by webReaderViewModel.annotationLiveData.observeAsState(null)
-  val shouldPopView: Boolean by webReaderViewModel.shouldPopViewLiveData.observeAsState(false)
-
-  val maxToolbarHeight = 48.dp
-  val maxToolbarHeightPx = with(LocalDensity.current) { maxToolbarHeight.roundToPx().toFloat() }
-  val toolbarHeightPx = remember { mutableStateOf(maxToolbarHeightPx) }
-
-  // Create a connection to the nested scroll system and listen to the scroll happening inside child Column
-  val nestedScrollConnection = remember {
-    object : NestedScrollConnection {
-      override fun onPreScroll(available: Offset, source: NestedScrollSource): Offset {
-        val delta = available.y
-        val newHeight = toolbarHeightPx.value + delta
-        toolbarHeightPx.value = newHeight.coerceIn(0f, maxToolbarHeightPx)
-        return Offset.Zero
-      }
-    }
-  }
-
-  if (webReaderParams == null) {
-    webReaderViewModel.loadItem(slug = slug)
-  }
-
-  if (webReaderParams != null) {
-    Box(
-      modifier = Modifier
-        .fillMaxSize()
-        .nestedScroll(nestedScrollConnection)
-    ) {
-      Column(
-        modifier = Modifier
-          .fillMaxSize()
-          .verticalScroll(webReaderViewModel.scrollState)
-
-      ) {
-        Row(
-          modifier = Modifier
-            .fillMaxWidth()
-            .requiredHeight(height = maxToolbarHeight)
-        ) {
-        }
-        WebReader(webReaderParams!!, webReaderViewModel.storedWebPreferences(isSystemInDarkTheme()), webReaderViewModel)
-      }
-
-      TopAppBar(
-        modifier = Modifier
-          .height(height = with(LocalDensity.current) {
-            webReaderViewModel.currentToolbarHeight = toolbarHeightPx.value.toInt()
-            toolbarHeightPx.value.roundToInt().toDp()
-          } ),
-        backgroundColor = MaterialTheme.colorScheme.surfaceVariant,
-        title = {},
-        actions = {
-          // Disabling menu until we implement local persistence
-          IconButton(onClick = { isMenuExpanded = true }) {
-            Icon(
-              imageVector = Icons.Filled.Menu,
-              contentDescription = null
-            )
-          }
-          IconButton(onClick = { showWebPreferencesDialog = true }) {
-            Icon(
-              imageVector = Icons.Filled.Settings, // TODO: set a better icon
-              contentDescription = null
-            )
-          }
-          SavedItemContextMenu(
-            isExpanded = isMenuExpanded,
-            isArchived = webReaderParams!!.item.isArchived,
-            onDismiss = { isMenuExpanded = false },
-            actionHandler = { webReaderViewModel.handleSavedItemAction(webReaderParams!!.item.savedItemId, it) }
-          )
-        }
-      )
-
-      if (showWebPreferencesDialog) {
-        WebPreferencesDialog(
-          onDismiss = {
-            showWebPreferencesDialog = false
-          },
-          webReaderViewModel = webReaderViewModel
-        )
-      }
-
-      if (annotation != null) {
-        AnnotationEditView(
-          initialAnnotation = annotation!!,
-          onSave = {
-            webReaderViewModel.saveAnnotation(it)
-          },
-          onCancel = {
-            webReaderViewModel.cancelAnnotationEdit()
-          }
-        )
-      }
-    }
-
-    LaunchedEffect(shouldPopView) {
-      if (shouldPopView) {
-        onBackPressedDispatcher?.onBackPressed()
-      }
-    }
-  } else {
-    // TODO: add a proper loading view
-    Text("Loading...")
-  }
-}
-
 @SuppressLint("SetJavaScriptEnabled")
 @Composable
 fun WebReader(
