+++ android/Omnivore/app/src/main/java/app/omnivore/omnivore/ui/save/SaveContent.kt
@@ -1,5 +1,7 @@
 package app.omnivore.omnivore.ui.save
 
+import android.content.Intent
+import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.MaterialTheme
@@ -11,13 +13,18 @@ import androidx.compose.material.*
 import androidx.compose.material.ButtonDefaults
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.unit.dp
+import app.omnivore.omnivore.MainActivity
+import app.omnivore.omnivore.ui.reader.PDFReaderActivity
+import app.omnivore.omnivore.ui.reader.WebReaderLoadingContainerActivity
 import kotlinx.coroutines.launch
 
 @Composable
 @OptIn(ExperimentalMaterialApi::class)
 fun SaveContent(viewModel: SaveViewModel, modalBottomSheetState: ModalBottomSheetState, modifier: Modifier) {
   val coroutineScope = rememberCoroutineScope()
+  val context = LocalContext.current
 
   Surface(modifier = Modifier.fillMaxSize(), color = MaterialTheme.colors.background) {
     Column(
@@ -34,7 +41,11 @@ fun SaveContent(viewModel: SaveViewModel, modalBottomSheetState: ModalBottomShee
           onClick = {
             coroutineScope.launch {
               modalBottomSheetState.hide()
-              // TODO: open app
+              viewModel.clientRequestID?.let {
+                val intent = Intent(context, WebReaderLoadingContainerActivity::class.java)
+                intent.putExtra("SAVED_ITEM_REQUEST_ID", it)
+                context.startActivity(intent)
+              }
             }
           },
           colors = ButtonDefaults.buttonColors(
